refactor: request url을 path, queryString으로 파싱하는 기능 분리
