+++ tomcat/src/test/java/org/utils/RequestLineParserTest.java
@@ -2,6 +2,7 @@ package org.utils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.common.request.parser.RequestLineParser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
